MNT: resolve UI search date at "day" granularity
The previous end datetime was 12:00AM on the date of app launch, which
excludes any entries created that day.  By removing the time from the
edit box and setting the end datetime filter arg to the end of the day,
time filters include any entries made on any day within the range.

Time granularity can still be used via the API, but not in the UI.